requirements.txt
Flask==2.0.1
Flask-CORS==3.0.10
pytesseract==0.3.10
Pillow==9.5.0
gunicorn==20.1.0 